# Improve `fold_at_level` performance